pyproject.toml
[tool.poetry]
name = "improved_test_code"
version = "0.1.0"
description = ""
authors = ["yomapi <yomapi.new@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.2"
PyYAML = "^6.0"
mysqlclient = "^2.1.1"
black = "^22.10.0"
djangorestframework = "^3.14.0"
pytest-django = "^4.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
